Added files API

+++ internal/storage/sqlite_crud.go
@@ -13,7 +13,7 @@ func (service *learningStore) CreateTable() error {
 		CREATE TABLE IF NOT EXISTS users (
 			id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
 			username TEXT UNIQUE,
-			password_hash TEXT,
+			password TEXT,
 			salt BLOB,  
 			first_name TEXT,
 			last_name TEXT,
@@ -46,11 +46,13 @@ func (service *learningStore) CreateTable() error {
 		CREATE TABLE IF NOT EXISTS learning_entries (
 			id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
 			goal_id INTEGER,
+			user_id INTEGER,
 			title TEXT,
 			description TEXT,
 			date DATETIME DEFAULT CURRENT_TIMESTAMP,
 			status TEXT DEFAULT 'Not Started',
 			FOREIGN KEY (goal_id) REFERENCES learning_goals(id)
+			FOREIGN KEY (user_id) REFERENCES users(id)
 		)
 	`)
 	if err != nil {
@@ -61,15 +63,15 @@ func (service *learningStore) CreateTable() error {
 	_, err = service.DB.Exec(`
 		CREATE TABLE IF NOT EXISTS learning_files (
 			id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-			goal_id INTEGER,
+			entry_id INTEGER,
+			user_id INTEGER,
 			filename TEXT,
 			filesize INTEGER,
 			filetype TEXT,
 			filepath TEXT,
 			created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-			owner_id INTEGER,
-			FOREIGN KEY (goal_id) REFERENCES learning_goals(id),
-			FOREIGN KEY (owner_id) REFERENCES users(id)
+			FOREIGN KEY (entry_id) REFERENCES learning_entries(id),
+			FOREIGN KEY (user_id) REFERENCES users(id)
 		)
 	`)
 	if err != nil {
@@ -81,10 +83,10 @@ func (service *learningStore) CreateTable() error {
 
 // User CRUD Methods
 // CreateUser create a user's details in the database
-func (service *learningStore) CreateUser(username string, passwordHash string, hash []byte, firstName string, lastName string) (int64, error) {
+func (service *learningStore) CreateUser(username string, password []byte, salt []byte, firstName string, lastName string) (int64, error) {
 	result, err := service.DB.Exec(`
-		INSERT INTO users (username, password_hash, hash, first_name, last_name, role) VALUES (?, ?, ?, ?, ?)
-	`, username, passwordHash, hash, firstName, lastName)
+		INSERT INTO users (username, password, salt, first_name, last_name) VALUES (?, ?, ?, ?, ?)
+	`, username, password, salt, firstName, lastName)
 
 	if err != nil {
 		return 0, err
@@ -111,8 +113,8 @@ func (service *learningStore) GetUserByID(id int) (models.User, error) {
 func (service *learningStore) GetUserByUsername(username string) (models.User, error) {
 	var user models.User
 	err := service.DB.QueryRow(`
-		SELECT id, username, first_name, last_name, created_at, updated_at, role FROM users WHERE username=?
-	`, username).Scan(&user.ID, &user.Username, &user.FirstName, &user.LastName, &user.CreatedAt, &user.UpdatedAt, &user.Role)
+		SELECT id, password, salt FROM users WHERE username=?
+	`, username).Scan(&user.ID, &user.Password, &user.Salt)
 
 	if err != nil {
 		return user, err
@@ -213,10 +215,10 @@ func (service *learningStore) GetGoalByID(id int) (models.LearningGoals, error) 
 }
 
 // CreateEntry inserts a new learning entry into the database and returns its ID.
-func (service *learningStore) CreateEntry(goalID int, title string, description string) (int64, error) {
+func (service *learningStore) CreateEntry(goalID int, userID int, title string, description string) (int64, error) {
 	result, err := service.DB.Exec(`
-        INSERT INTO learning_entries (goal_id, title, description) VALUES (?, ?, ?)
-    `, goalID, title, description)
+        INSERT INTO learning_entries (goal_id, user_id,  title, description) VALUES (?, ?, ?, ?)
+    `, goalID, userID, title, description)
 	if err != nil {
 		return 0, err
 	}
@@ -225,10 +227,10 @@ func (service *learningStore) CreateEntry(goalID int, title string, description 
 }
 
 // UpdateEntry updates an existing learning entry.
-func (service *learningStore) UpdateEntry(id int, title string, description string, date time.Time, status string) error {
+func (service *learningStore) UpdateEntry(id int, goal_id int, user_id int, title string, description string, date time.Time, status string) error {
 	_, err := service.DB.Exec(`
-        UPDATE learning_entries SET title=?, description=?, date=?, status=? WHERE id=?
-    `, title, description, date, status, id)
+        UPDATE learning_entries SET title=?, description=?, date=?, status=? WHERE id=? and goal_id=? and user_id=?
+    `, title, description, date, status, id, goal_id, user_id)
 	if err != nil {
 		return err
 	}
@@ -236,10 +238,10 @@ func (service *learningStore) UpdateEntry(id int, title string, description stri
 }
 
 // DeleteEntry deletes a learning entry by ID.
-func (service *learningStore) DeleteEntry(id int) error {
+func (service *learningStore) DeleteEntry(id int, userID int) error {
 	_, err := service.DB.Exec(`
-        DELETE FROM learning_entries WHERE id=?
-    `, id)
+        DELETE FROM learning_entries WHERE id=? and user_id=?
+    `, id, userID)
 	if err != nil {
 		return err
 	}
@@ -250,7 +252,7 @@ func (service *learningStore) DeleteEntry(id int) error {
 func (service *learningStore) GetAllEntriesByGoalID(goalID int) ([]models.LearningEntry, error) {
 	var entries []models.LearningEntry
 	rows, err := service.DB.Query(`
-        SELECT id, goal_id, title, description, date, status FROM learning_entries WHERE goal_id=?
+        SELECT id, goal_id, user_id, title, description, date, status FROM learning_entries WHERE goal_id=?
     `, goalID)
 	if err != nil {
 		return entries, err
@@ -259,7 +261,7 @@ func (service *learningStore) GetAllEntriesByGoalID(goalID int) ([]models.Learni
 
 	for rows.Next() {
 		var entry models.LearningEntry
-		err = rows.Scan(&entry.ID, &entry.LearningGoalID, &entry.Title, &entry.Description, &entry.Date, &entry.Status)
+		err = rows.Scan(&entry.ID, &entry.GoalID, &entry.UserID, &entry.Title, &entry.Description, &entry.Date, &entry.Status)
 		if err != nil {
 			return entries, err
 		}
@@ -272,8 +274,8 @@ func (service *learningStore) GetAllEntriesByGoalID(goalID int) ([]models.Learni
 func (service *learningStore) GetEntryByID(id int) (models.LearningEntry, error) {
 	var entry models.LearningEntry
 	err := service.DB.QueryRow(`
-        SELECT id, goal_id, title, description, date, status FROM learning_entries WHERE id=?
-    `, id).Scan(&entry.ID, &entry.LearningGoalID, &entry.Title, &entry.Description, &entry.Date, &entry.Status)
+        SELECT id, goal_id, user_id, title, description, date, status FROM learning_entries WHERE id=?
+    `, id).Scan(&entry.ID, &entry.GoalID, &entry.UserID, &entry.Title, &entry.Description, &entry.Date, &entry.Status)
 	if err != nil {
 		return entry, err
 	}
@@ -281,10 +283,10 @@ func (service *learningStore) GetEntryByID(id int) (models.LearningEntry, error)
 }
 
 // CreateFile inserts a new learning file into the database and returns its ID.
-func (service *learningStore) CreateFile(goalID int, ownerID int, fileName string, fileSize int64, fileType string, filePath string) (int64, error) {
+func (service *learningStore) CreateFile(entryID int, userID int, fileName string, fileSize int64, fileType string, filePath string) (int64, error) {
 	result, err := service.DB.Exec(`
-        INSERT INTO learning_files (goal_id, owner_id, filename, filesize, filetype, filepath) VALUES (?, ?, ?, ?, ?, ?)
-    `, goalID, ownerID, fileName, fileSize, fileType, filePath)
+        INSERT INTO learning_files (entry_id, user_id, filename, filesize, filetype, filepath) VALUES (?, ?, ?, ?, ?, ?)
+    `, entryID, userID, fileName, fileSize, fileType, filePath)
 	if err != nil {
 		return 0, err
 	}
@@ -293,10 +295,10 @@ func (service *learningStore) CreateFile(goalID int, ownerID int, fileName strin
 }
 
 // UpdateFile updates an existing learning file.
-func (service *learningStore) UpdateFile(id int, goalID int, ownerID int, fileName string, fileSize int64, fileType string) error {
+func (service *learningStore) UpdateFile(id int, entryID int, userID int, fileName string, fileSize int64, fileType string) error {
 	_, err := service.DB.Exec(`
-        UPDATE learning_files SET goal_id=?, owner_id=?, filename=?, filesize=?, filetype=? WHERE id=?
-    `, goalID, ownerID, fileName, fileSize, fileType, id)
+        UPDATE learning_files SET entry_id=?, user_id=?, filename=?, filesize=?, filetype=? WHERE id=?
+    `, entryID, userID, fileName, fileSize, fileType, id)
 	if err != nil {
 		return err
 	}
@@ -315,10 +317,10 @@ func (service *learningStore) DeleteFile(id int) error {
 }
 
 // GetAllFilesByGoalID returns all learning files for a given goal ID.
-func (service *learningStore) GetAllFilesByGoalID(goalID int) ([]models.LearningFiles, error) {
+func (service *learningStore) GetAllFilesByEntryID(goalID int) ([]models.LearningFiles, error) {
 	var files []models.LearningFiles
 	rows, err := service.DB.Query(`
-        SELECT id, goal_id, owner_id, filename, filesize, filetype, filepath FROM learning_files WHERE goal_id=?
+        SELECT id, entry_id, user_id, filename, filesize, filetype, filepath FROM learning_files WHERE entry_id=?
     `, goalID)
 	if err != nil {
 		return files, err
@@ -327,7 +329,7 @@ func (service *learningStore) GetAllFilesByGoalID(goalID int) ([]models.Learning
 
 	for rows.Next() {
 		var file models.LearningFiles
-		err = rows.Scan(&file.ID, &file.LearningGoalID, &file.OwnerID, &file.FileName, &file.FileSize, &file.FileType, &file.FilePath)
+		err = rows.Scan(&file.ID, &file.EntryID, &file.UserID, &file.FileName, &file.FileSize, &file.FileType, &file.FilePath)
 		if err != nil {
 			return files, err
 		}
@@ -340,8 +342,8 @@ func (service *learningStore) GetAllFilesByGoalID(goalID int) ([]models.Learning
 func (service *learningStore) GetFileByID(id int) (models.LearningFiles, error) {
 	var file models.LearningFiles
 	err := service.DB.QueryRow(`
-        SELECT id, goal_id, owner_id, filename, filesize, filetype, filepath FROM learning_files WHERE id=?
-    `, id).Scan(&file.ID, &file.LearningGoalID, &file.OwnerID, &file.FileName, &file.FileSize, &file.FileType, &file.FilePath)
+        SELECT id, entry_id, user_id, filename, filesize, filetype, filepath FROM learning_files WHERE id=?
+    `, id).Scan(&file.ID, &file.EntryID, &file.UserID, &file.FileName, &file.FileSize, &file.FileType, &file.FilePath)
 	if err != nil {
 		return file, err
 	}
